moved many stuffs out of headers
