+++ README.md
@@ -1,1 +1,8 @@
-# TODO: a *lot* of refactoring 
+# TODO:
+- figure out how to `make` multiple at once?
+- remove ancientCrap.h
+- remove easyAttr.h
+- autoescape quotes in makefile
+- remove `HEADER_OKAY`
+- rename `USAGE_STRING`
+- remove list's `*tail`
