+++ testing/lab4Pract.c
@@ -1,7 +1,5 @@
-// NB: you can have multiple lines in your USAGE_STRING
-#define USAGE_STRING "inputFile outputFile chunks\nwhere chunks >= 1"
-#define ERR_MULTIPROCESS 0
 #include "katwikOpsys.h"
+//USAGE_STRING="inputFile outputFile chunks\nwhere chunks >= 1"
 
 #include <sys/types.h>
 
@@ -11,22 +9,21 @@ int main(int argc, char** argv) {
 	USAGE(argc == 4);
 	char* inputFile = argv[1];
 	char* outputFile = argv[2];
+
 	int chunks;
 	USAGE(( chunks = atoi(argv[3]) ) >= 1);
 
-	int inputFildes, outputFildes;
-	ERR_NEG1(inputFildes = open(inputFile, O_RDONLY));
-	ERR_NEG1(outputFildes = open(outputFile, O_WRONLY));
+	int inputFildes = open_(inputFile, O_RDONLY),
+		outputFildes = open_(outputFile, O_WRONLY);
 
-	char* readBuffer;
-	ERR_NULL(readBuffer = malloc(chunks * BUF_SIZE * sizeof(char)));
+	char* readBuffer = malloc_(chunks * BUF_SIZE * sizeof(char));
 	parallelIO(NULL, inputFildes, 2, BUF_SIZE, (void*) readBuffer, NULL, chunks, READ);
 
 	printBuf(readBuffer, chunks * BUF_SIZE);
 	FREE(readBuffer);
 
-	ERR_NEG1(close(inputFildes));
-	ERR_NEG1(close(outputFildes));
+	close_(inputFildes);
+	close_(outputFildes);
 
 	return EXIT_SUCCESS;
 }
